pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-project"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.2",
    "openpyxl>=3.1",
    "sqlalchemy>=2.0",
    "PyMySQL>=1.1",
    "PyYAML>=6.0",
    "pyarrow"
]